app connected to node web socket server.
next up: showing a push notification on device screen as a result of something happening between client and server.

so maybe the client sends the server a message, the server receives, does something with it and sends a message back and it the client receives it as a push notification?